did you know we had pulse grenade noises w enever used (#410)
* Update emgrenade.dm

* Update emgrenade.dm

+++ code/game/objects/items/grenades/emgrenade.dm
@@ -1,5 +1,5 @@
 /obj/item/grenade/empgrenade
-	name = "classic EMP grenade"
+	name = "pulse grenade"
 	desc = "It is designed to wreak havoc on electronic systems."
 	icon_state = "emp"
 	item_state = "emp"
@@ -7,5 +7,9 @@
 /obj/item/grenade/empgrenade/prime(mob/living/lanced_by)
 	. = ..()
 	update_mob()
+	var/empgrenade_turf = get_turf(src)
+	if(!empgrenade_turf)
+		return
+	playsound(empgrenade_turf, 'sound/f13weapons/pulsegrenade.ogg', 100, TRUE, 8, 0.9)
 	empulse_using_range(src, 14)
 	qdel(src)
